% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendExtraPageBlocks.R
\name{appendExtraPageBlocks}
\alias{appendExtraPageBlocks}
\title{appendExtraPageBlocks}
\usage{
appendExtraPageBlocks(blocks, add)
}
\arguments{
\item{blocks}{a nested list for extra page blocks per page}

\item{add}{a seccond nested list for extra page blocks per page to be
appended to the first one}
}
\description{
A helper to merge two nested lists describing extra page blocks.
The lists have the page name on the first level and flattened documentation
blocks on the second level. It is ensured that elements for the
same page are grouped in the same list.
}
\author{
Falk Benke
}
