% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildHTML.R
\name{buildHTML}
\alias{buildHTML}
\title{buildHTML}
\usage{
buildHTML(
  style = "classic",
  folder = "html",
  mdfolder = "markdown",
  literature = "literature.bib",
  citation = "../CITATION.cff",
  supplementary = "images",
  debug = FALSE,
  templatefolder = ".."
)
}
\arguments{
\item{style}{visualization style to be used for the creation. Currently available styles are 
"classic" and "ming"}

\item{folder}{location the HTML files should be written to}

\item{mdfolder}{path to the markdown folder to be used as source}

\item{literature}{path to a bibliography, if available (will be ignored
if file does not exist)}

\item{citation}{Citation information in citation file format (optional)}

\item{supplementary}{a vector of files and/or folders required for the conversion
(e.g. an images subdirectory with figures to be shown in the documents)}

\item{debug}{logical which switches on/off a debug mode which will return additional 
status updates and keep build files}

\item{templatefolder}{Folder in which goxygen will search for template files in addition to the pre-installed ones.}
}
\description{
Converts a folder with markdown files and a corresponding literature library 
(if available) to HTML files and creates cross-links between them.
}
\details{
Pandoc (https://pandoc.org/) together with pandoc-citeproc need to be installed 
on the system.
}
\seealso{
\code{\link{goxygen}}, \code{\link{buildTEX}}
}
\author{
Jan Philipp Dietrich
}
