% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMarkdown.R
\name{buildMarkdown}
\alias{buildMarkdown}
\title{buildMarkdown}
\usage{
buildMarkdown(x, folder = "markdown")
}
\arguments{
\item{x}{a named list of markdown codes which should be written as markdown files. 
The name of each entry will become the file name.}

\item{folder}{folder the markdown files should be written to}
}
\description{
Creates a folder filled with markdown files from a list object with markdown code
}
\seealso{
\code{\link{goxygen}}, \code{\link{buildHTML}}
}
\author{
Jan Philipp Dietrich
}
