% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pandoc.R
\name{check_pandoc}
\alias{check_pandoc}
\title{check_pandoc}
\usage{
check_pandoc()
}
\description{
Support function which checks pandoc availability and stops with an error in case
that pandoc cannot be found
}
\author{
Jan Philipp Dietrich
}
