% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseTemplate.R
\name{chooseTemplate}
\alias{chooseTemplate}
\title{chooseTemplate}
\usage{
chooseTemplate(style, templatefolder, ftype)
}
\arguments{
\item{style}{visualization style to be used for the creation.}

\item{templatefolder}{Folder in which goxygen will search for template files in addition to the pre-installed ones.}

\item{ftype}{template file type / file ending}
}
\description{
Support function helping to choose the selected template
}
\seealso{
\code{\link{goxygen}}, \code{\link{buildTEX}}
}
\author{
Jan Philipp Dietrich
}
