% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createListModularCode.R
\name{createListModularCode}
\alias{createListModularCode}
\title{createListModularCode}
\usage{
createListModularCode(
  cc,
  interfaces,
  path = ".",
  citation = NULL,
  unitPattern = c("\\\\(", "\\\\)"),
  includeCore = FALSE,
  mainfile = "main.gms",
  docfolder = "doc",
  startType = "equations"
)
}
\arguments{
\item{cc}{codeCheck information}

\item{interfaces}{interface information}

\item{path}{path to the model to be documented}

\item{citation}{citation data read from a CFF file}

\item{unitPattern}{pattern that is usedto identify the unit in the description, default =c("\\(","\\)")}

\item{includeCore}{Boolean whether core should be included or not, default=FALSE}

\item{mainfile}{main file of the model}

\item{docfolder}{folder the documentation should be written to relative to model folder}

\item{startType}{input parameter for \code{\link{extractDocumentation}} to be passed
when extracting documentation from realizations. Defaults to "equations",
meaning that documentation in realizations should be interpreted as equations
documentation, if no identifier is set.}
}
\description{
support function to create documentation of modular GAMS code.
}
\seealso{
\code{\link[gms]{codeCheck}}
}
\author{
Jan Philipp Dietrich
}
