% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createListSimpleCode.R
\name{createListSimpleCode}
\alias{createListSimpleCode}
\title{createListSimpleCode}
\usage{
createListSimpleCode(path = ".", citation = NULL, mainfile = "main.gms")
}
\arguments{
\item{path}{path to the model to be documented}

\item{citation}{citation data read from a CFF file}

\item{mainfile}{main file of the model}
}
\description{
support function to create documentation of non-modular GAMS code.
}
\seealso{
\code{\link{codeCheck}}
}
\author{
Jan Philipp Dietrich
}
