% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createModulePage.R
\name{createModulePage}
\alias{createModulePage}
\title{createModulePage}
\usage{
createModulePage(data, docfolder)
}
\arguments{
\item{data}{a list of data entries for the resulting markdown page. Following 
entries can be provided:
\describe{
  \item{name}{Name of the module}
  \item{title}{Page title}
  \item{description}{General description}
  \item{input}{Table containing inputs to the module}
  \item{output}{Table containing outputs from the module}
  \item{realizations}{A list of realizations with entries
  "description" and "limitations" for each of them}
  \item{declarations}{Table of declarations for internal objects}
  \item{stes}{Table containing sets used in the module}
  \item{authors}{Module authors}
  \item{seealso}{A vector with names of relevant other documentation pages.}
}}

\item{docfolder}{folder the documentation should be written to relative to model folder}
}
\value{
a character vector containing the produced markdown text
}
\description{
Creates markdown code from a supplied data list
}
\seealso{
\code{\link{goxygen}}
}
\author{
Jan Philipp Dietrich
}
