% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSimplePage.R
\name{createSimplePage}
\alias{createSimplePage}
\title{createSimplePage}
\usage{
createSimplePage(data)
}
\arguments{
\item{data}{a list of data entries for the resulting markdown page. Following 
entries can be provided:
\describe{
  \item{title}{Page title}
  \item{description}{General description}
  \item{limitations}{Limitations the implementation comes with}
  \item{authors}{Module authors}
}}
}
\value{
a character vector containing the produced markdown text
}
\description{
Creates markdown code from a supplied data list
}
\seealso{
\code{\link{goxygen}}
}
\author{
Jan Philipp Dietrich
}
