% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.empty}
\alias{.empty}
\title{.empty}
\usage{
.empty(zz)
}
\arguments{
\item{zz}{a connection object of class "textConnection" containing the markdown document}
}
\description{
helper function which adds an empty line in a markdown document
}
\seealso{
\code{\link{goxygen}}, \code{\link{createModulePage}}
}
\author{
Jan Philipp Dietrich
}
