% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.header}
\alias{.header}
\title{.header}
\usage{
.header(zz, title, level, id = NULL)
}
\arguments{
\item{zz}{a connection object of class "textConnection" containing the markdown document}

\item{title}{the title to be used (character vector of length 1)}

\item{level}{level of the heading (1 means main header, higher numbers reflect lower levels)}

\item{id}{ID given to the title (relevant for anchors)}
}
\description{
helper function which writes a title for a markdown section
}
\seealso{
\code{\link{goxygen}}, \code{\link{createModulePage}}
}
\author{
Jan Philipp Dietrich
}
