% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.limitations}
\alias{.limitations}
\title{.limitations}
\usage{
.limitations(zz, limitations, emptyIfNULL = FALSE)
}
\arguments{
\item{zz}{a connection object of class "textConnection" containing the markdown document}

\item{limitations}{A character vector containing the given limitations}

\item{emptyIfNULL}{switch which decides whether limitations section should be ignored, if
limitations input is NULL or if it should state that there are no known limitations.}
}
\description{
helper function which adds a "limitations" section.
}
\seealso{
\code{\link{goxygen}}, \code{\link{createModulePage}}
}
\author{
Jan Philipp Dietrich
}
