% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.section}
\alias{.section}
\title{.section}
\usage{
.section(data, zz, title, level, id = NULL)
}
\arguments{
\item{data}{a character vector to be written to the markdown document}

\item{zz}{a connection object of class "textConnection" containing the markdown document}

\item{title}{the title to be used (character vector of length 1)}

\item{level}{level of the heading (1 means main header, higher numbers reflect lower levels)}

\item{id}{ID given to the title (relevant for anchors)}
}
\description{
helper function which creates a section consisting of header and content in a markdown document
and skips section when content is empty
}
\seealso{
\code{\link{goxygen}}, \code{\link{createModulePage}}
}
\author{
Falk Benke
}
