% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.updateImagePaths}
\alias{.updateImagePaths}
\title{.updateImagePaths}
\usage{
.updateImagePaths(x)
}
\arguments{
\item{x}{A character vector containing image paths.}
}
\description{
helper function which updates relative image paths so that they refer to a subfolder
images instead of refering to the current folder.
}
\seealso{
\code{\link{goxygen}}, \code{\link{createModulePage}}
}
\author{
Jan Philipp Dietrich
}
