% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.write}
\alias{.write}
\title{.write}
\usage{
.write(zz, data)
}
\arguments{
\item{zz}{a connection object of class "textConnection" containing the markdown document}

\item{data}{a character vector to be written to the markdown document}
}
\description{
helper function which writes a character vector line by line in a markdown document
}
\seealso{
\code{\link{goxygen}}, \code{\link{createModulePage}}
}
\author{
Jan Philipp Dietrich
}
