% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenPageBlockList.R
\name{flattenPageBlockList}
\alias{flattenPageBlockList}
\title{flattenPageBlockList}
\usage{
flattenPageBlockList(data)
}
\arguments{
\item{data}{a list of documentation pieces with type as name of each element}
}
\value{
a list with two elements (1) `blocks` containing the documentation elements
with type as name of the element and (2) `extraPageBlocks` containing lists for
blocks to be put on an extra pages, sorted by page names
}
\description{
A helper that processes additional attributes for a given list of code documentation
blocks. Code documentation blocks are described as lists consisting of `content`
containing the documentation and a `cfg` list containing attributes.
}
\details{
Supports sublists with code documentation per realization.#'
}
\author{
Falk Benke
}
