% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamsequation2tex.R
\name{gamsequation2tex}
\alias{gamsequation2tex}
\title{gamsequation2tex}
\usage{
gamsequation2tex(x)
}
\arguments{
\item{x}{GAMS equation provided as character}
}
\value{
GAMS equation converted to latex code
}
\description{
Convert a gams equation into latex code
}
\examples{

  x <- "eq_1 .. v_a =e= sum(j,v_b(j)*((1-s_c)+sum(cell(i,j),v_d(i)/f_d(i))));"
  cat(gamsequation2tex(x))
}
\seealso{
\code{\link{goxygen}}
}
\author{
Jan Philipp Dietrich
}
