% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldBuildHTML.R
\name{oldBuildHTML}
\alias{oldBuildHTML}
\title{oldBuildHTML}
\usage{
oldBuildHTML(
  folder = "html",
  mdfolder = "markdown",
  literature = "literature.bib",
  citation = "../CITATION.cff",
  supplementary = NULL,
  addHTML = NULL
)
}
\arguments{
\item{folder}{location the HTML files should be written to}

\item{mdfolder}{path to the markdown folder to be used as source}

\item{literature}{path to a bibliography, if available (will be ignored
if file does not exist)}

\item{citation}{Citation information in citation file format (optional)}

\item{supplementary}{a vector of files and/or folders required for the conversion
(e.g. an images subdirectory with figures to be shown in the documents)}

\item{addHTML}{character vector with HTML code which should be added to the body of each HTML file.}
}
\description{
Converts a folder with markdown files and a corresponding literature library 
(if available) to HTML files and creates cross-links between them.
}
\details{
Pandoc (https://pandoc.org/) together with pandoc-citeproc need to be installed 
on the system.
}
\seealso{
\code{\link{goxygen}}, \code{\link{buildTEX}}
}
\author{
Jan Philipp Dietrich
}
