% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnReferences.R
\name{returnReferences}
\alias{returnReferences}
\title{returnReferences}
\usage{
returnReferences(names, targets, file, level = 2)
}
\arguments{
\item{names}{vector of reference names}

\item{targets}{vector of reference adresses (same order and lengths as names)}

\item{file}{name of the reference file to be written}

\item{level}{level of the "References" title to be written}
}
\description{
Support function to create a reference file linking references with corresponding adresses.
}
\author{
Jan Philipp Dietrich
}
