% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortExtraPages.R
\name{sortExtraPages}
\alias{sortExtraPages}
\title{sortExtraPages}
\usage{
sortExtraPages(extraPage)
}
\arguments{
\item{extraPage}{flat list of blocks with names indicating identifier and
extrapage (e.g. 'description-settings')}
}
\description{
helper transforming a flat list of extra page blocks into
sublists per page
}
\author{
Falk Benke
}
