% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTEX.R
\name{buildTEX}
\alias{buildTEX}
\title{buildTEX}
\usage{
buildTEX(
  file = "documentation.tex",
  mdfolder = "markdown",
  literature = "literature.bib",
  citation = "../CITATION.cff",
  supplementary = NULL,
  pdf = TRUE,
  style = "classic",
  templatefolder = ".."
)
}
\arguments{
\item{file}{name of the tex file to be written}

\item{mdfolder}{path to the markdown folder to be used as source}

\item{literature}{path to a bibliography, if available (will be ignored
if file does not exist)}

\item{citation}{Citation information in citation file format (optional)}

\item{supplementary}{a vector of files and/or folders required for the conversion
(e.g. an images subdirectory with figures to be shown in the documents)}

\item{pdf}{boolean which specifies whether pdf file should be generated from tex}

\item{style}{visualization style to be used for the Latex/PDF creation. Currently only "classic" style is 
available. Ignored for outputs other than Latex/PDF. Can be extended by additional templates stored in the
\code{templatefolder} in the format \code{<style>.latex}. Classic template 
\code{system.file("templates","classic.latex",package="goxygen")} can serve as a starting point for own 
templates.}

\item{templatefolder}{Folder in which goxygen will search for template files in addition to the pre-installed ones.}
}
\description{
Converts a folder with markdown files and a corresponding literature library 
(if available) to a tex file
}
\details{
Pandoc (https://pandoc.org/) together with pandoc-citeproc need to be installed 
on the system.
}
\seealso{
\code{\link{goxygen}}, \code{\link{buildHTML}}
}
\author{
Jan Philipp Dietrich, Kristine Karstens
}
